refactor: configures registry client auth during client creation
A new repository is created each time a remote method is called because
the source reference may not be in the same repository each call. Currently, the
authentication client is being recreated on these calls as well, but this is
not required so this change creates the authentication client during registry
client instantiation.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>